Move error conversions to separate file
